+++ db/migrate/20110804115347_devise_create_users.rb
@@ -1,6 +1,6 @@
 class DeviseCreateUsers < ActiveRecord::Migration
   def self.up
-    create_table(:users) do |t|
+    create_table(:users) unless table_exists? :users do |t|
       t.database_authenticatable :null => false
       t.recoverable
       t.rememberable
@@ -14,13 +14,13 @@ class DeviseCreateUsers < ActiveRecord::Migration
       t.timestamps
     end
 
-    add_index :users, :email,                :unique => true
-    add_index :users, :reset_password_token, :unique => true
+    add_index :users, :email,                :unique => true unless index_exists? :users, :email
+    add_index :users, :reset_password_token, :unique => true unless index_exists? :users, :reset_password_token
     # add_index :users, :confirmation_token,   :unique => true
     # add_index :users, :unlock_token,         :unique => true
   end
 
   def self.down
-    drop_table :users
+    drop_table :users if table_exists? :users
   end
 end
