+++ db/migrate/20120123053346_add_download_notes.rb
@@ -1,9 +1,9 @@
 class AddDownloadNotes < ActiveRecord::Migration
   def self.up
-    add_column :data_source_downloads, :notes, :text
+    add_column :data_source_downloads, :notes, :text unless column_exists? :data_source_downloads, :notes
   end
 
   def self.down
-    remove_column :data_source_downloads, :notes
+    remove_column :data_source_downloads, :notes if column_exists? :data_source_downloads, :notes
   end
 end
