+++ db/migrate/20160419203702_add_reset_password_token_and_reset_password_sent_at_to_users.rb
@@ -1,9 +1,9 @@
 class AddResetPasswordTokenAndResetPasswordSentAtToUsers < ActiveRecord::Migration
   def self.up
-    add_column :users, :reset_password_sent_at, :datetime
+    add_column :users, :reset_password_sent_at, :datetime unless column_exists? :users, :reset_password_sent_at
   end
 
   def self.down
-    remove_column :users, :reset_password_sent_at
+    remove_column :users, :reset_password_sent_at if column_exists? :users, :reset_password_sent_at
   end
 end
